Expose SingleFileLoggerFactory as the public interface
