+++ pulsar-client-cpp/include/pulsar/Client.h
@@ -28,6 +28,8 @@
 #include <pulsar/MessageBuilder.h>
 #include <pulsar/ClientConfiguration.h>
 #include <pulsar/Schema.h>
+#include <pulsar/SimpleLoggerFactory.h>
+#include <pulsar/SingleFileLoggerFactory.h>
 #include <string>
 
 namespace pulsar {
